Utils update
-saveData method moved
